requirements.txt
asgiref==3.8.1
Django==5.0.14
djangorestframework==3.16.0
mssql-django==1.5
pillow==11.2.1
pyodbc==5.2.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
